+++ app/src/main/java/com/techcourse/controller/mvc/MvcRegisterController.java
@@ -2,11 +2,23 @@ package com.techcourse.controller.mvc;
 
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
+import context.org.springframework.stereotype.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import web.org.springframework.web.bind.annotation.RequestMapping;
+import web.org.springframework.web.bind.annotation.RequestMethod;
+import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.mvc.asis.CustomController;
+import webmvc.org.springframework.web.servlet.view.JspView;
 
+@Controller
 public class MvcRegisterController implements CustomController {
+
+    @RequestMapping(value = "/register", method = RequestMethod.POST)
+    public ModelAndView save(HttpServletRequest req, HttpServletResponse res) throws Exception {
+        final String path = execute(req, res);
+        return new ModelAndView(new JspView(path));
+    }
 
     @Override
     public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
