--- app/src/main/java/com/techcourse/controller/mvc/MvcRegisterViewController.java
@@ -1,13 +1,0 @@
-package com.techcourse.controller.mvc;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.mvc.asis.CustomController;
-
-public class MvcRegisterViewController implements CustomController {
-
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        return "/register.jsp";
-    }
-}
