+++ app/src/main/java/com/techcourse/controller/mvc/view/MvcLoginController.java
@@ -1,17 +1,28 @@
-package com.techcourse.controller.mvc;
+package com.techcourse.controller.mvc.view;
 
+import com.techcourse.controller.mvc.MvcUserSession;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
+import context.org.springframework.stereotype.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import web.org.springframework.web.bind.annotation.RequestMapping;
+import web.org.springframework.web.bind.annotation.RequestMethod;
+import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.mvc.asis.CustomController;
+import webmvc.org.springframework.web.servlet.view.JspView;
 
-@context.org.springframework.stereotype.Controller
+@Controller
 public class MvcLoginController implements CustomController {
 
     private static final Logger log = LoggerFactory.getLogger(MvcLoginController.class);
+
+    @RequestMapping(value = "/login", method = RequestMethod.GET)
+    public ModelAndView login(HttpServletRequest req, HttpServletResponse res) throws Exception {
+        return new ModelAndView(new JspView(execute(req, res)));
+    }
 
     @Override
     public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
